refactor(syscall): widen the syscall ABI to pointer-width (uintptr_t)
The SVC syscall ABI passed args and returned results as uint32_t, mirroring the
ARM register width. That silently can't carry a pointer on a 64-bit host: a
pointer arg truncates to 32 bits on the way in (e.g. SYS_open's path, SYS_write's
buffer), and a pointer result truncates on the way out (SYS_malloc). It is a
latent defect in the ABI, and the specific blocker to running SVC on the host
port.

Make args uintptr_t and the return intptr_t — pointer-width by definition. On
ARMv7-M uintptr_t == uint32_t, so the svc register ABI, the exception frame, and
the codegen are byte-identical; the asm SVCall_Handler is typeless and unchanged.
On a 64-bit host a pointer now round-trips losslessly, so no MAP_32BIT low-heap
constraint and no -m32 multilib are needed to carry pointers through a syscall.

Touched: the v_svc0..3 trampolines (portable/cortex-m4/port_syscall.h + the host
model in tests/stubs/port_syscall.h), v_syscall_dispatch and its malloc/calloc/
realloc pointer returns (portable/cortex-m4/syscall.c), v_host_svc
(tests/stubs/syscall_stubs.c), and the call sites in ipc.c/devfs.c/memory.c that
dropped their truncating (uint32_t)(uintptr_t) pointer casts. Unrelated 32-bit
pointer casts (the utils.c PSP overflow check) were left untouched.

Verified: ARM firmware clean under NAVHAL=OFF and NAVHAL=ON; an ARM SVC build
(FD_IPC_TEST) compiles the widened v_svc asm cleanly; host unit suite 246/246
including the syscall-dispatch and fd-typed-IPC tests that exercise the dispatch;
host demo runs; tripwire green. Closes most of the deferred 32-bit-pointer-ABI
item and is the foundation for host SVC + per-task region enforcement.

+++ kernel/devfs.c
@@ -112,7 +112,7 @@ void v_fd_close_all(TCB *t) {
 int v_file_open(const char *path, int flags) {
 #if VAIOS_SYSCALL_SVC
   if (v_in_thread_mode())
-    return v_svc2(SYS_open, (uint32_t)(uintptr_t)path, (uint32_t)flags);
+    return v_svc2(SYS_open, (uintptr_t)path, (uint32_t)flags);
 #endif
   (void)flags;
   dev_node_t *node = dev_find(path);
@@ -124,7 +124,7 @@ int v_file_open(const char *path, int flags) {
 int v_file_write(int fd, const void *buf, uint32_t len) {
 #if VAIOS_SYSCALL_SVC
   if (v_in_thread_mode())
-    return v_svc3(SYS_write, (uint32_t)fd, (uint32_t)(uintptr_t)buf, len);
+    return v_svc3(SYS_write, (uint32_t)fd, (uintptr_t)buf, len);
 #endif
   v_fd_entry *e = fd_lookup(fd);
   if (!e || !e->ops->write)
@@ -135,7 +135,7 @@ int v_file_write(int fd, const void *buf, uint32_t len) {
 int v_file_read(int fd, void *buf, uint32_t len) {
 #if VAIOS_SYSCALL_SVC
   if (v_in_thread_mode())
-    return v_svc3(SYS_read, (uint32_t)fd, (uint32_t)(uintptr_t)buf, len);
+    return v_svc3(SYS_read, (uint32_t)fd, (uintptr_t)buf, len);
 #endif
   v_fd_entry *e = fd_lookup(fd);
   if (!e || !e->ops->read)
